backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
torch
transformers
scipy
openai-whisper
soundfile
numpy
librosa
sentencepiece
datasets
